+++ src/main/java/com/selesse/jxlint/model/ProgramOptions.java
@@ -1,9 +1,10 @@
 package com.selesse.jxlint.model;
 
+import com.google.common.base.Joiner;
 import com.google.common.base.Splitter;
 import com.google.common.collect.Lists;
 import com.google.common.collect.Maps;
-import com.selesse.jxlint.ProgramExitter;
+import com.selesse.jxlint.model.rules.Categories;
 import com.selesse.jxlint.model.rules.LintRulesImpl;
 import com.selesse.jxlint.model.rules.NonExistentLintRuleException;
 
@@ -86,20 +87,41 @@ public class ProgramOptions {
      * Returns a list of strings from the raw option string. In other words,
      * given a raw option string "Rule1, Rule2" (i.e. parsed from <code>--disable "Rule1, Rule2"</code>), this returns
      * a list with { "Rule1", "Rule2" }. If any of the rules passed as strings do not exist, this will function
-     * will exit the program.
+     * will throw a {@link com.selesse.jxlint.model.rules.NonExistentLintRuleException}.
      */
-    public static List<String> getListFromRawOptionStringOrDie(String disabledRules) {
+    public static List<String> getRuleListFromOptionString(String optionString)
+            throws NonExistentLintRuleException {
         Splitter splitter = Splitter.on(",").omitEmptyStrings().trimResults();
 
-        List<String> disabledRulesStringList = Lists.newArrayList(splitter.split(disabledRules));
-        for (String disabledRuleString : disabledRulesStringList) {
-            try {
-                LintRulesImpl.getInstance().getLintRule(disabledRuleString);
-            } catch (NonExistentLintRuleException e) {
-                ProgramExitter.exitProgramWithMessage(e.getMessage(), ExitType.COMMAND_LINE_ERROR);
-            }
+        List<String> rulesStringList = Lists.newArrayList(splitter.split(optionString));
+        for (String disabledRuleString : rulesStringList) {
+            LintRulesImpl.getInstance().getLintRule(disabledRuleString);
         }
 
-        return Lists.newArrayList(disabledRulesStringList);
+        return Lists.newArrayList(rulesStringList);
+    }
+
+    public static List<String> getCategoryListFromOptionString(String categoryOptionString)
+            throws IllegalArgumentException {
+        List<String> categoryList = Lists.newArrayList();
+
+        Splitter splitter = Splitter.on(",").omitEmptyStrings().trimResults();
+        List<String> rawCategoryStringList = Lists.newArrayList(splitter.split(categoryOptionString));
+
+        Enum[] categories = Categories.get().getEnumConstants();
+        List<String> categoryNames = Lists.newArrayList();
+        for (Enum category : categories) {
+            categoryNames.add(category.toString());
+        }
+
+        for (String categoryString : rawCategoryStringList) {
+            if (!categoryNames.contains(categoryString)) {
+                throw new IllegalArgumentException("Category \"" + categoryString + "\" does not exist. Try one of: " +
+                        Joiner.on(", ").join(categories) + ".");
+            }
+            categoryList.add(categoryString);
+        }
+
+        return categoryList;
     }
 }