+++ src/test/java/com/selesse/jxlint/DispatcherTest.java
@@ -1,13 +1,13 @@
 package com.selesse.jxlint;
 
-import com.google.common.base.Charsets;
 import com.google.common.base.Joiner;
 import com.google.common.collect.Lists;
 import com.google.common.io.Files;
-import com.google.common.io.Resources;
 import com.selesse.jxlint.linter.Linter;
 import com.selesse.jxlint.linter.LinterFactory;
 import com.selesse.jxlint.model.ExitType;
+import com.selesse.jxlint.model.rules.Category;
+import com.selesse.jxlint.model.rules.LintError;
 import com.selesse.jxlint.model.rules.LintRulesImpl;
 import com.selesse.jxlint.samplerules.xml.XmlLintRulesTestImpl;
 import com.selesse.jxlint.settings.JxlintProgramSettings;
@@ -17,7 +17,6 @@ import org.junit.Test;
 
 import java.io.File;
 import java.io.IOException;
-import java.net.URISyntaxException;
 import java.util.List;
 
 import static org.junit.Assert.assertEquals;
@@ -38,32 +37,34 @@ public class DispatcherTest extends AbstractTestCase {
     public void testHelpProperlyExtracted() {
         String newLineSeparator = System.getProperty("line.separator");
 
-        final List<String> expectedOutput = Lists.newArrayList(
-                "usage: jxlint [flags] <directory>"
-                , " -h,--help                Usage information, help message."
-                , " -v,--version             Output version information."
-                , " -p,--profile             Measure time every rule takes to complete."
-                , " -l,--list                Lists lint rules with a short, summary"
-                , "                          explanation."
-                , " -r,--rules               Prints a Markdown file containing the program's"
-                , "                          rules."
-                , " -s,--show <RULE[s]>      Lists a verbose rule explanation."
-                , " -c,--check <RULE[s]>     Only check for these rules."
-                , " -d,--disable <RULE[s]>   Disable the list of rules."
-                , " -e,--enable <RULE[s]>    Enable the list of rules."
-                , " -w,--nowarn              Only check for errors; ignore warnings."
-                , " -Wall,--Wall             Check all warnings, including those off by"
-                , "                          default."
-                , " -Werror,--Werror         Treat all warnings as errors."
-                , " -q,--quiet               Don't output any progress or reports."
-                , " -t,--html <filename>     Create an HTML report."
-                , " -x,--xml <filename>      Create an XML (!!) report."
-                , ""
-                , "<RULE[s]> should be comma separated, without spaces."
-                , "Exit Status:"
-                , "0                     Success"
-                , "1                     Failed"
-                , "2                     Command line error" + newLineSeparator
+        List<String> expectedOutput = Lists.newArrayList(
+                "usage: jxlint [flags] <directory>",
+                " -h,--help                     Usage information, help message.",
+                " -v,--version                  Output version information.",
+                " -p,--profile                  Measure time every rule takes to complete.",
+                " -l,--list                     Lists lint rules with a short, summary",
+                "                               explanation.",
+                " -r,--rules                    Prints a Markdown dump of the program's",
+                "                               rules.",
+                " -s,--show <RULE[s]>           Lists a verbose rule explanation.",
+                " -c,--check <RULE[s]>          Only check for these rules.",
+                " -d,--disable <RULE[s]>        Disable the list of rules.",
+                " -e,--enable <RULE[s]>         Enable the list of rules.",
+                " -y,--category <CATEGORY[s]>   Run all rules of a certain category.",
+                " -w,--nowarn                   Only check for errors; ignore warnings.",
+                " -Wall,--Wall                  Check all warnings, including those off by",
+                "                               default.",
+                " -Werror,--Werror              Treat all warnings as errors.",
+                " -q,--quiet                    Don't output any progress or reports.",
+                " -t,--html <filename>          Create an HTML report.",
+                " -x,--xml <filename>           Create an XML (!!) report.",
+                "",
+                "<RULE[s]> should be comma separated, without spaces.",
+                "Exit Status:",
+                "0                     Success",
+                "1                     Failed",
+                "2                     Command line error",
+                ""
         );
 
         runExitTest(new String[] { "--help" }, tempDirectory, Joiner.on(newLineSeparator).join(expectedOutput), ExitType.SUCCESS);
@@ -76,86 +77,6 @@ public class DispatcherTest extends AbstractTestCase {
                 programSettings.getProgramVersion());
 
         runExitTest(new String[] { "--version" }, tempDirectory, expectedOutput, ExitType.SUCCESS);
-    }
-
-    @Test
-    public void testMarkdownReportProperlyGenerated() throws URISyntaxException, IOException {
-        File markdownDump = new File(Resources.getResource("markdown-dump.md").toURI());
-        List<String> markdownDumpContents = Files.readLines(markdownDump, Charsets.UTF_8);
-        String expectedOutput = Joiner.on("\n").join(markdownDumpContents);
-        ProgramSettings programSettings = new JxlintProgramSettings();
-        expectedOutput = expectedOutput.replace("$$$VERSION$$$", programSettings.getProgramVersion());
-
-        runExitTest(new String[]{"-r"}, tempDirectory, expectedOutput, ExitType.SUCCESS);
-    }
-
-    @Test
-    public void testShowInvalidOptionSaysSo() {
-        final String expectedOutput = "'foobarrule' is not a valid rule.";
-
-        runExitTest(new String[] { "--show=foobarrule" }, tempDirectory, expectedOutput, ExitType.COMMAND_LINE_ERROR);
-    }
-
-    @Test
-    public void testShowInvalidOptionSaysSoVariation() {
-        final String expectedOutput = "'foobarrule' is not a valid rule.";
-
-        runExitTest(new String[] { "--show", "foobarrule" }, tempDirectory, expectedOutput,
-                ExitType.COMMAND_LINE_ERROR);
-    }
-
-    @Test
-    public void testShowValidOption() {
-        final List<String> expectedOutput = Lists.newArrayList(
-                  "XML encoding specified"
-                , "----------------------"
-                , "Summary: Encoding of the XML must be specified."
-                , ""
-                , "Severity: Warning"
-                , "Category: LINT"
-                , ""
-                , "The XML encoding should be specified. For example, <?xml version=\"1.0\" encoding=\"UTF-8\"?>.\n\n"
-        );
-
-        runExitTest(new String[] { "--show", "XML encoding specified" }, tempDirectory, Joiner.on("\n").join(expectedOutput),
-                ExitType.SUCCESS);
-    }
-
-    @Test
-    public void testShowValidDisabledByDefaultOption() {
-        String expectedRuleOutput =
-                "Author tag specified\n" +
-                "--------------------\n" +
-                "Summary: author.xml files must contain a valid root-element <author> tag.\n" +
-                "\n" +
-                "** Disabled by default **\n" +
-                "\n" +
-                "Severity: Warning\n" +
-                "Category: STYLE\n" +
-                "\n" +
-                "For style purposes, every author.xml file must contain an <author> tag as the\n" +
-                "root element. This tag should also have the 'name' and 'creationDate'\n" +
-                "attributes. For example:\n" +
-                "\n" +
-                "    <?xml version=\"1.0\" encoding=\"UTF-8\">\n" +
-                "    <author name=\"Steve Holt\" creationDate=\"2013-09-28\">\n" +
-                "      .. content ..\n" +
-                "    </author>\n\n";
-
-        runExitTest(new String[] { "--show", "Author tag specified" }, tempDirectory,
-                expectedRuleOutput, ExitType.SUCCESS);
-    }
-
-    @Test
-    public void testListOptions() {
-        final List<String> expectedOutput = Lists.newArrayList(
-               "\"Unique attribute\" : Attributes within a tag must be unique.",
-               "\"XML version specified\" : Version of XML must be specified.",
-               "\"XML encoding specified\" : Encoding of the XML must be specified.",
-               "\"Author tag specified\"* : author.xml files must contain a valid root-element <author> tag."
-        );
-
-        runExitTest(new String[] { "--list" }, tempDirectory, Joiner.on("\n").join(expectedOutput), ExitType.SUCCESS);
     }
 
     @Test
@@ -292,5 +213,32 @@ public class DispatcherTest extends AbstractTestCase {
         setupTestLinterAndRunProgramWithArgs(new String[] { "--nowarn", tempDirectory.getAbsolutePath() });
         linter = LinterFactory.getInstance();
         assertEquals(0, linter.getLintErrors().size());
+    }
+
+    @Test
+    public void testCategoryReturnsRightRules() {
+        TestFileCreator.createBadAuthorFile(tempDirectory);
+        TestFileCreator.createBadVersionFile(tempDirectory);
+        TestFileCreator.createBadEncodingFile(tempDirectory);
+        TestFileCreator.createBadAttributeFile(tempDirectory);
+
+        setupTestLinterAndRunProgramWithArgs(new String[]{"--category", Category.LINT.toString(),
+                tempDirectory.getAbsolutePath()});
+
+        Linter linter = LinterFactory.getInstance();
+        assertEquals(5, linter.getLintErrors().size());
+
+        for (LintError lintError : linter.getLintErrors()) {
+            assertEquals(Category.LINT, lintError.getViolatedRule().getCategory());
+        }
+
+        setupTestLinterAndRunProgramWithArgs(new String[]{"--category", Category.PERFORMANCE.toString(),
+                tempDirectory.getAbsolutePath()});
+        linter = LinterFactory.getInstance();
+        assertEquals(2, linter.getLintErrors().size());
+
+        for (LintError lintError : linter.getLintErrors()) {
+            assertEquals(Category.PERFORMANCE, lintError.getViolatedRule().getCategory());
+        }
     }
 }